# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Akmatori API
  description: |
    AI-powered AIOps platform API. Manages skills, tools, incidents, alert sources,
    and system settings for automated incident investigation and remediation.
  version: 1.0.0
  contact:
    name: Akmatori
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
...
    AggregationSettings:
      type: object
      properties:
        id:
          type: integer
        enabled:
          type: boolean
        correlation_confidence_threshold:
          type: number
        merge_confidence_threshold:
          type: number
        recorrelation_enabled:
          type: boolean
        recorrelation_interval_minutes:
          type: integer
        max_incidents_to_analyze:
          type: integer
        observing_duration_minutes:
          type: integer
        correlator_timeout_seconds:
          type: integer
        merge_analyzer_timeout_seconds:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /settings/proxy:
    get:
      summary: Get proxy settings
      operationId: getProxySettings
      tags: [Settings]
      responses:
        '200':
          description: Proxy configuration
    put:
      summary: Update proxy settings
      operationId: updateProxySettings
      tags: [Settings]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                proxy_url:
                  type: string
                  format: uri
                no_proxy:
                  type: string
                services:
                  type: object
                  properties:
                    openai:
                      type: object
                      properties:
                        enabled:
                          type: boolean
                    slack:
                      type: object
                      properties:
                        enabled:
                          type: boolean
                    zabbix:
                      type: object
                      properties:
                        enabled:
                          type: boolean
      responses:
        '200':
          description: Updated proxy settings
  /settings/aggregation:
    get:
      summary: Get aggregation settings
      operationId: getAggregationSettings
      tags: [Settings]
      responses:
        '200':
          description: Aggregation settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregationSettings'
    put:
      summary: Update aggregation settings
      operationId: updateAggregationSettings
      tags: [Settings]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AggregationSettings'
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregationSettings'

  # ===== Context Files =====
...